Needs cleanup, Fix flaw in A* which invalidated all searches after the first. The best, shortest path is now evaluated from pathfinding results, and not chosen based on shortest dist between perimeter points. Cleanup is necessary
Since it's a static map, I decided to not regenerate the nodes in
between searches since they wont change. However, since the search data
was stored in the nodes, the search was dirtied with data from previous
searches. This was fixed by decoupling the search data from the nodes
into the open set and only giving it a lifetime of the current search.